refactor: move anr tests to correct package
